Move MaxMind class to its own module
